requirements.txt
numpy==1.21.4
pickle-mixin==1.0.2
streamli
scikit-learn==1.0.1
protobuf == 3.20.1
